refactor: created SubmitButton component
I have created a submit button component because I have applied style changes with the help of hook
There are multiple forms which uses this submit button (login,registration). In the previous implementation
I have created the style inside the form components but now the styling happens inside the submit button component.
